Quotation marks.
Tested a few different methods in creating the quotation marks like in
the mockup. However, the typography of the resulting marks could not be
matched. Will discuss in class in Friday. Also, unable to figure out
how to italicize the “per 100 g” above the detailed list. Will also
discuss this in class.