Found solutions to quotations and italics.
Designed quotation marks in the same font family as the headers, which
does not match the mockup but it still provides the same stylized feel
from the mockup. Also, placed <i> phrasing element around the “per 100
g” and used the tag to cover the bases and stylized it in css.